setuid for binary removed as default value
The default value for the mode of the beuser binary should not include
the setuid bit (for root, by default). If the module is applied without
any parameters set, it would allow any normal user to change to any other.
Instead, the default is that nobody may switch.
The preferred way of controlling who may call beuser is via sudo.
If needed, setuid can still be enabled by a actively changing the defaults.